System.h declutter 2 - Move PWM channel allocation to LedcPin.cpp

+++ Grbl_Esp32/src/Spindles/PWMSpindle.h
@@ -60,7 +60,7 @@ namespace Spindles {
 
     protected:
         int32_t  _current_pwm_duty;
-        uint8_t  _pwm_chan_num;
+        int      _pwm_chan_num;
         uint32_t _pwm_period;     // how many counts in 1 period
         uint8_t  _pwm_precision;  // auto calculated
